+++ src/GafferSceneUI/LightTool.cpp
@@ -47,6 +47,7 @@
 #include "GafferUI/ImageGadget.h"
 #include "GafferUI/StandardStyle.h"
 
+#include "Gaffer/Animation.h"
 #include "Gaffer/Metadata.h"
 #include "Gaffer/MetadataAlgo.h"
 #include "Gaffer/NameValuePlug.h"
@@ -72,8 +73,14 @@
 #include "boost/algorithm/string/predicate.hpp"
 #include "boost/bind/bind.hpp"
 
+#include "OpenEXR/OpenEXRConfig.h"
+#if OPENEXR_VERSION_MAJOR < 3
+#include "OpenEXR/ImathMatrixAlgo.h"
+#include "OpenEXR/ImathSphere.h"
+#else
 #include "Imath/ImathMatrixAlgo.h"
 #include "Imath/ImathSphere.h"
+#endif
 
 #include "fmt/format.h"
 
@@ -142,6 +149,19 @@ Plug *activeValuePlug( Plug *sourcePlug )
 		return tweakPlug->valuePlug();
 	}
 	return sourcePlug;
+}
+
+void setValueOrAddKey( FloatPlug *plug, float time, float value )
+{
+	if( Animation::isAnimated( plug ) )
+	{
+		Animation::CurvePlug *curve = Animation::acquire( plug );
+		curve->insertKey( time, value );
+	}
+	else
+	{
+		plug->setValue( value );
+	}
 }
 
 const char *constantFragSource()
@@ -791,7 +811,9 @@ class SpotLightHandle : public LightToolHandle
 					}
 
 					// Clamp each individual cone angle as well
-					coneFloatPlug->setValue(
+					setValueOrAddKey(
+						coneFloatPlug,
+						m_view->getContext()->getTime(),
 						conePlugAngle(
 							clampHandleAngle(
 								originalConeHandleAngle + angleDelta,
@@ -812,7 +834,9 @@ class SpotLightHandle : public LightToolHandle
 					}
 
 					// Clamp each individual cone angle as well
-					penumbraFloatPlug->setValue(
+					setValueOrAddKey(
+						penumbraFloatPlug,
+						m_view->getContext()->getTime(),
 						penumbraPlugAngle(
 							clampHandleAngle(
 								originalPenumbraHandleAngle.value() + angleDelta,