+++ Changes.md
@@ -9,6 +9,14 @@ Features
 --------
 
 - USDShader : Added a node for loading shaders from USD's `SdrRegistry`. This includes shaders such as `UsdPreviewSurface` and `UsdUVTexture`, which are now available in the `USD/Shader` section of the node menu.
+- USDLight : Added a node for defining UsdLux lights. This is available from the `USD/Light` section of the node menu.
+
+Improvements
+------------
+
+- LightEditor : Added section displaying UsdLux shadow parameters.
+- Cycles : Added support for UsdLux lights.
+- LightTool : Added support for editing animated plugs.
 
 Fixes
 -----
@@ -18,6 +26,10 @@ Fixes
 - Arnold
   - Fixed translation of `vector` typed outputs defined as `vector <name>` in an output definition.
   - Fixed translation of `shadow:enable` and `shadow:color` parameters on UsdLux lights, which were previously ignored.
+- LightTool : Fixed bug causing non-settable plugs to be enabled, such as plugs with an expression as input or those using the default spreadsheet row for their value. This generated the error `ERROR : EventSignalCombiner : Cannot set value for plug {plug} except during computation`.
+- LightEditor : Fixed toggling values in cases where inherited light attributes were set by a script context variable without including a default.
+- GLWidget : Fixed rare crash when showing a GLWidget for the first time.
+- BranchCreator : Fixed bug which could cause inconsistent hashes to be generated.
 
 API
 ---
@@ -248,6 +260,8 @@ Fixes
 - Arnold
   - Fixed translation of `vector` typed outputs defined as `vector <name>` in an output definition.
   - Fixed translation of `shadow:enable` and `shadow:color` parameters on UsdLux lights, which were previously ignored.
+- BranchCreator : Fixed bug which could cause inconsistent hashes to be generated.
+- LightEditor : Fixed toggling values in cases where inherited light attributes were set by a script context variable without including a default.
 
 1.2.10.0 (relative to 1.2.9.0)
 ========